refactor: move link widget to internal/ui package and remove redundant text
Move newLinkWithCopy to internal/ui.NewLinkWithCopy so it's a proper
reusable package outside of cmd/. Remove the "✓" text feedback since
the icon swap to ConfirmIcon already provides sufficient visual
indication through the button's built-in styling.
