+++ internal/ui/link_widget.go
@@ -1,4 +1,4 @@
-package main
+package ui
 
 import (
 	"net/url"
@@ -12,10 +12,10 @@ import (
 
 const copyFeedbackDuration = 1500 * time.Millisecond
 
-// newLinkWithCopy creates a hyperlink with a small copy-to-clipboard button beside it.
+// NewLinkWithCopy creates a hyperlink with a small copy-to-clipboard button beside it.
 // The link is displayed using Fyne's Hyperlink widget (underlined, themed color).
 // The copy button shows a clipboard icon and briefly switches to a checkmark on tap.
-func newLinkWithCopy(text, rawURL string, w fyne.Window) fyne.CanvasObject {
+func NewLinkWithCopy(text, rawURL string, w fyne.Window) fyne.CanvasObject {
 	parsedURL, _ := url.Parse(rawURL)
 
 	link := widget.NewHyperlink(text, parsedURL)
@@ -26,11 +26,9 @@ func newLinkWithCopy(text, rawURL string, w fyne.Window) fyne.CanvasObject {
 		w.Clipboard().SetContent(rawURL)
 
 		copyBtn.SetIcon(theme.ConfirmIcon())
-		copyBtn.SetText("✓")
 
 		time.AfterFunc(copyFeedbackDuration, func() {
 			fyne.Do(func() {
-				copyBtn.SetText("")
 				copyBtn.SetIcon(theme.ContentCopyIcon())
 			})
 		})
